talayie/requirements/base.txt
psycopg2==2.9.3
python-dotenv
loguru
alembic==1.11.0
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
dominate==2.7.0
email-validator==2.0.0.post2
Faker==18.10.1
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
visitor==0.1.3
Werkzeug==2.3.4
WTForms==3.0.1
flask-restx